#56 問題点の修正
